Move thread logic from comment service
