+++ backend/tests/test_thread_service/test_get_threads_heads.py
@@ -6,7 +6,7 @@ def test_get_threads_heads():
     thread_service.create_thread(content="", topic="chaos")
     thread_service.create_thread(content="", topic="watch")
 
-    heads = comment_service.get_threads_heads()
+    heads = thread_service.get_threads_heads()
     assert len(heads) == 3
 
 
@@ -14,10 +14,10 @@ def test_only_threads_heads():
     thread_service.create_thread(content="", topic="it")
     comment_service.create_comment(content="", parent_id=1)
 
-    heads = comment_service.get_threads_heads()
+    heads = thread_service.get_threads_heads()
     assert len(heads) == 1
 
 
 def test_no_threads_heads():
-    heads = comment_service.get_threads_heads()
+    heads = thread_service.get_threads_heads()
     assert len(heads) == 0
