+++ backend/tests/test_thread_service/test_get_topic_threads_heads.py
@@ -6,7 +6,7 @@ def test_get_topic_threads_heads():
     thread_service.create_thread(content="", topic="it")
     thread_service.create_thread(content="", topic="it")
 
-    heads = comment_service.get_topic_threads_heads(topic="it")
+    heads = thread_service.get_topic_threads_heads(topic="it")
     for head in heads:
         assert head.topic == "it"
         assert head.parent_id is None
@@ -17,7 +17,7 @@ def test_exact_thread_topic():
     thread_service.create_thread(content="", topic="watch")
     thread_service.create_thread(content="", topic="politics")
 
-    heads = comment_service.get_topic_threads_heads(topic="it")
+    heads = thread_service.get_topic_threads_heads(topic="it")
     assert len(heads) == 1
     assert heads[0].topic == "it"
 
@@ -27,6 +27,6 @@ def test_no_children():
     comment_service.create_comment(content="child", parent_id=1)
     comment_service.create_comment(content="child", parent_id=2)
 
-    heads = comment_service.get_topic_threads_heads(topic="it")
+    heads = thread_service.get_topic_threads_heads(topic="it")
     assert len(heads) == 1
     assert heads[0].content == "head"
